program for ticket counter
You have to choose a data structure to hold the details of a movie-ticket holder. Following are the details:
Name of the purchaser, Phone number, Movie's name, Time, Date.
The person issuing the tickets at the counter, issues the tickets on a 'First come First Serve basis'.
Write a function 'issue_ticket()' to handle issuing the tickets (say 10 tickets) and another function ' tickets_available()' to count the number of available tickets at any time.